requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
asyncio==3.4.3
psutil==5.9.6
netifaces==0.11.0
python-iptables==1.0.0 